review/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.3.0
python-dotenv==1.0.1
httpx==0.24.1
PyPDF2==3.0.1
numpy==1.26.0
arxiv
pyyaml
